requirements.txt
celery>=4.2.1
Django>=2.1
django-allauth>=0.36.0
django-appconf>=1.0.2
django_celery_beat>=1.1.1
django-graphos-3>=1.1
django-import-export>=1.1.0
django-sekizai>=0.10.0
jsonfield>=2.0.2
pandas>=0.23.3
pandas-datareader>=0.6.0
tiingo>=0.9.0
pytz>=2018.5
stripe>=2.8.0
virtualenv>=16.0.0
pycryptodome>=3.6.6
numpy>=1.14.5
TA-Lib==0.4.17
boto3>=1.9.79
quandl>=3.4.0
sklearn>=0.0
pycurl==7.43.0.1
matplotlib>=3.0.3
mpl-finance>=0.10.0
statsmodels>=0.9.0